Add ingest package
