+++ java/system-test/system-test-suite/src/main/java/sleeper/systemtest/suite/dsl/ingest/SystemTestDirectIngest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.systemtest.suite.dsl;
+package sleeper.systemtest.suite.dsl.ingest;
 
 import sleeper.core.record.Record;
 import sleeper.systemtest.datageneration.GenerateNumberedRecords;
